added CustomerController, removed WeatherController, Small change in demo client app to use new endpoint
